Feat: 리사이즈 파일 관리를 위한 커스텀 훅 제작
